pyproject.toml
[tool.poetry]
name = "bookrag"
version = "0.1.0"
description = ""
authors = ["Marat <soaptradep@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
langchain = "^0.3.11"
langchain-google-genai = "^2.0.6"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.11"
sentence-transformers = "^3.3.1"
chromadb = "^0.5.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
